fixing dippindots script

+++ dippindots.sh
@@ -162,13 +162,22 @@ cp "./assets/Inconsolata-dz.otf" ~/Library/Fonts
 
 # =============== XVIM =================================
 tput setaf 5
-echo "Installing XVim (Vim for XCode)..."
+read -p "Do you want install Vim bindings for XCode (XVim)? (y/n) " -n 1
 tput sgr0
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+	tput setaf 5
+	echo "Installing XVim (Vim for XCode)..."
+	tput sgr0
 
-# Install XVim (Vim for XCode)
-git clone https://github.com/JugglerShu/XVim.git
-xcodebuild -project XVim/XVim.xcodeproj
-rm -rf XVim
+	# Install XVim (Vim for XCode)
+	git clone https://github.com/JugglerShu/XVim.git
+	xcodebuild -project XVim/XVim.xcodeproj
+	rm -rf XVim
+else
+	tput setaf 3
+	echo "Skipping XVim..."
+	tput sgr0
+fi
